requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
hashids==0.8.4
pytz==2021.3
sqlparse==0.4.2
uWSGI==2.0.15ß
